Move get_docstring from module_formatter into ansible/utils
